feat(viewer): replace the link-share banner with an author origin ⓘ
Anonymous link-share pages no longer carry a persistent disclaimer bar.
A small ⓘ next to the author opens a popover (desktop) or bottom sheet
(phone) that explains the page was shared by an Artifact Share user and
offers the existing report dialog. The report API, events, cooldowns,
and Slack notification are unchanged.
