+++ apps/web/app/guides/link-sharing.ja.md
@@ -11,7 +11,7 @@
 
 受け取り用のURLでは、ファイルごとに専用のサブドメインを使います。所有者がログインして管理するときは、引き続き`https://artifactshare.com/a/<id>`を開きます。
 
-受け取り側の画面には、利用者が共有した内容であることを示す案内と「通報」が常に表示されます。新しいアカウントが共有した内容からArtifact Share外のリンクをクリックした場合は、移動先を示す確認画面で続行またはキャンセルを選びます（修飾キー付きで新しいタブに開く操作では表示されません）。
+受け取り側の画面では、作成者名の横の小さな ⓘ から、Artifact Shareの利用者が共有したページであることの説明と「通報」を開けます。新しいアカウントが共有した内容からArtifact Share外のリンクをクリックした場合は、移動先を示す確認画面で続行またはキャンセルを選びます（修飾キー付きで新しいタブに開く操作では表示されません）。
 
 選べる終了日は、ワークスペースに設定された最長期限までです。新しいリンクにはワークスペースの既定期限があらかじめ入ります。既定は30日、最長は90日です。オーナーまたは管理者が設定を変更している場合は、その内容が優先されます。
