+++ apps/web/app/i18n/en.json
@@ -1337,7 +1337,9 @@
   "vw.sandboxPaused.next": "You can resume right away.",
   "vw.sandboxPaused.retryFailed": "The last attempt did not finish. Try again or reload the page.",
   "vw.sandboxResuming": "Resuming display…",
-  "vw.linkSafety.disclaimer": "This content was shared by an Artifact Share user. Artifact Share has not reviewed it.",
+  "vw.linkSafety.disclaimer": "This page was published by an Artifact Share user. Artifact Share has not reviewed or endorsed its content.",
+  "vw.linkOrigin.open": "About this page",
+  "vw.linkOrigin.title": "About this page",
   "vw.linkSafety.report": "Report",
   "vw.linkSafety.reportTitle": "Report this content",
   "vw.linkSafety.reportDescription": "Tell us why this shared content may be unsafe.",
