+++ apps/web/app/i18n/ja.json
@@ -1337,7 +1337,9 @@
   "vw.sandboxPaused.next": "すぐに表示を再開できます。",
   "vw.sandboxPaused.retryFailed": "直前の再開処理を完了できませんでした。もう一度試すか、ページを再読み込みしてください。",
   "vw.sandboxResuming": "表示を再開しています…",
-  "vw.linkSafety.disclaimer": "この内容は Artifact Share の利用者が共有したもので、Artifact Share が内容を確認したものではありません。",
+  "vw.linkSafety.disclaimer": "このページは Artifact Share の利用者が共有したもので、Artifact Share が内容を確認・保証したものではありません。",
+  "vw.linkOrigin.open": "このページについて",
+  "vw.linkOrigin.title": "このページについて",
   "vw.linkSafety.report": "通報",
   "vw.linkSafety.reportTitle": "この内容を通報",
   "vw.linkSafety.reportDescription": "共有された内容に問題があると思われる理由を選んでください。",
